requirements.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
redis==5.0.1
asgiref==3.7.2
